Deprecate old references to properties serializer/deserializer.
