+++ core/serialization/deserializer/properties-deserializer.js
@@ -1,55 +1,3 @@
-var Montage = require("../../core").Montage;
+require("../../deprecate").deprecationWarning("montage/core/serialization/deserializer/properties-deserializer", "montage/core/serialization/deserializer/values-deserializer");
 
-var PropertiesDeserializer = Montage.specialize( {
-    _object: {value: null},
-    _objectDescriptor: {value: null},
-    _context: {value: null},
-
-    initWithObjectAndObjectDescriptorAndContext: {
-        value: function (object, objectDescriptor, context) {
-            this._object = object;
-            this._objectDescriptor = objectDescriptor;
-            this._context = context;
-
-            return this;
-        }
-    },
-
-    get: {
-        value: function (name) {
-            if (this._objectDescriptor.values) {
-                return this._objectDescriptor.values[name];
-            } else if (this._objectDescriptor.properties) { // deprecated
-                return this._objectDescriptor.properties[name];
-            }
-        }
-    },
-
-    deserializeProperties: {
-        value: function (propertyNames) {
-            var object = this._object,
-                // .properties deprecated
-                values = this._objectDescriptor.values || this._objectDescriptor.properties,
-                propertyName;
-
-            if (values) {
-                if (!propertyNames) {
-                    propertyNames = Montage.getSerializablePropertyNames(object);
-                }
-
-                for (var i = 0, ii = propertyNames.length; i < ii; i++) {
-                    propertyName = propertyNames[i];
-                    object[propertyName] = values[propertyName];
-                }
-            }
-        }
-    },
-
-    getObjectByLabel: {
-        value: function (label) {
-            this._context.getObject(label);
-        }
-    }
-});
-
-exports.PropertiesDeserializer = PropertiesDeserializer;
+exports.PropertiesDeserializer = require("./values-deserializer").ValuesDeserializer;
